requirements.txt
altgraph==0.17
click==7.1.2
cycler==0.10.0
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mysql-connector-python==8.0.23
numpy==1.19.5
pefile==2019.4.18
Pillow==8.1.0
protobuf==3.14.0
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
PyMySQL==1.0.2
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
six==1.15.0
Werkzeug==1.0.1
